miner/pyproject.toml
[project]
name = "caster-miner"
version = "0.1.0"
description = "Miner-facing SDK harness for the Caster Subnet."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
  { name = "Caster Labs", email = "engineering@casterlabs.example" }
]
dependencies = [
  "fastapi>=0.110.1",
  "httpx>=0.27",
  "python-dotenv>=1.1.1",
  "pyseccomp>=0.1.2",
  "uvicorn>=0.30",
  "bittensor==9.9.0",
  "caster-miner-sdk>=0.1.0",
]

[project.scripts]
caster-miner = "caster_miner.app:main"
caster-miner-dev = "caster_miner.dev:main"
caster-miner-submit = "caster_miner.submit:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/caster_miner"]
